fix(memory): cast fp32-kept memory_key_scale to key dtype in amp branch
P1 real-weight smoke caught it: from_pretrained keeps memory_key_scale
in fp32 (_keep_in_fp32_modules) while the model runs bf16; the amp mul
promoted the memory key slice (and via cat the whole key) to fp32 and
flash-attn rejected the q/k dtype mismatch. Tiny-model tests missed it
because .to(bf16) casts the scale too — regression test added with a
forced-fp32 scale on a bf16 model. Also move the CPU SDPA monkeypatch
to import time (unittest-discover module order made setUpModule
patching order-dependent) and add the P1 smoke script; rerun on real
Helios-Base weights: 13/13 PASS, peak 58.3 GiB.

+++ helios/modules/transformer_helios.py
@@ -425,7 +425,10 @@ class HeliosAttnProcessor:
                 )
 
         if not enable_cross and memory_context_length > 0 and getattr(attn, "is_amplify_memory", False):
-            scale_mem = attn.get_scale_memory().view(1, 1, -1, 1)
+            # memory_key_scale lives in _keep_in_fp32_modules: cast the scale to
+            # the key dtype, otherwise the mul promotes the slice (and cat the
+            # whole key) to fp32 and flash-attn rejects the q/k dtype mismatch.
+            scale_mem = attn.get_scale_memory().to(key.dtype).view(1, 1, -1, 1)
             key = torch.cat([key[:, :memory_context_length] * scale_mem, key[:, memory_context_length:]], dim=1)
 
         hidden_states = attn_varlen_func(
